refactor: Winners라는 레이싱 승자를 가려내기 위한 도메인을 추가하여 Cars의 책임 분리
