+++ src/main/java/racingcar/domain/Cars.java
@@ -42,20 +42,6 @@ public class Cars {
                 .collect(Collectors.toUnmodifiableList());
     }
 
-    public List<String> getWinnerCarsName() {
-        Car winner = getWinner();
-        return cars.stream()
-                .filter(car -> car.isSamePosition(winner))
-                .map(Car::getName)
-                .collect(Collectors.toUnmodifiableList());
-    }
-
-    private Car getWinner() {
-        return cars.stream()
-                .max(Car::compareTo)
-                .orElseThrow(() -> new IllegalArgumentException(ErrorConstant.ERROR_PREFIX + "비교할 자동차가 없습니다."));
-    }
-
     private void validateDuplicatedNames(List<String> carNames) {
         List<String> distinctCarNames = carNames.stream()
                 .distinct()
